RT75957. twr-k70f120m: use external memory for the malloc() pool
This is required for using the NAND Flash.

Since the generic U-Boot code requires more than 128 KB of RAM for the
Bad Block Table and we are not going to heavily modify the generic code,
we add optional support for the malloc() poll in the external memory in
order to have enough space for the Bad Block Table (the whole internal
SRAM is only 128 KB in size.) The CONFIG_SYS_MALLOC_EXT_BASE and the
CONFIG_SYS_MALLOC_EXT_LEN options control the base address and size of
the malloc() pool. If those preprocessor variables are not defined, use
the original code for determining the position of the malloc() pool in
memory.