RT77788. k70-som: Set DDR setting R2RSAME to 3 to avoid memory corruptions
 * When Ethernet and/or LCD are extensively utilized, the DDR controller send
command to the memory chip very often, so that sometimes these commands go
back-to-back to one another. In this situation the value of the R2RSAME
parameter is taken into account by the DDR controller.
 * The necessary value for CR37[R2RSAME] was found experimentally. The K70
Reference Manual is very unclear on the exact meanings of delays set by the
fields of the DDR_CR37 register (including R2RSAME).